Multiple bug fixes from switch to new LLVM-C APIs - Added LibLLVMDwarfAttributeEncoding enum - Dropped JIT handle interfaces infavor of simple ulong as the handle - Fixed Nuget.config for the native LibLLVMDwarfAttributeEncoding - Updated template generation to use the full package version in stead of the file version in header comments and attributes for better readability - Fixed ArrayMarhsalInfo to use pointee type as the element type for generated code - Added missing ArrayMarshalInfo entries to bindings generator config - Added a simple interop test to help find root cause of crashes in JIT - Fixed global handle release to correctly release the right handle value - Fixed bug in DebugInfoBuilder() constructor that incorrectly inverted the allowUnresolved parameter - Fixed bug in MDNode.ReplaceAllUsesWith as the now official LLVM API has slightly different behavior and deletes the node replaced - Simplified Enumeration of basic blocks, leveraging LLVM-C APIs for GetFirst/GetNext...

+++ src/Llvm.NET/Values/AttributeKindExtensions.cs
@@ -10,7 +10,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
 using Llvm.NET.Instructions;
-using Llvm.NET.Interop;
 using Llvm.NET.Properties;
 using Ubiquity.ArgValidators;
 
@@ -629,7 +628,7 @@ namespace Llvm.NET.Values
         {
             return ( from kind in Enum.GetValues( typeof( AttributeKind ) ).Cast<AttributeKind>( ).Skip( 1 )
                      let name = kind.GetAttributeName( )
-                     select new KeyValuePair<uint, AttributeKind>( LLVMGetEnumAttributeKindForName( name, ( size_t )name.Length ), kind )
+                     select new KeyValuePair<uint, AttributeKind>( LLVMGetEnumAttributeKindForName( name, name.Length ), kind )
                    ).ToDictionary( kvp => kvp.Key, kvp => kvp.Value );
         }
